Battle Zone discrete sound & discrete sound additions - added discrete sound to battle zone - engine frequency (resistor R11) can be changed with slider - discrete sound may be disabled by setting BZONE_DISCRETE to 0 in bzone.h
DISCRETE_RCDISC3
- added diode junction voltage 
- junction voltage also indicates polarity

DISCRETE_OPAMP_FILTER
- added type DISC_OP_AMP_FILTER_IS_LOW_PASS_1M
- INP2 takes role of v_ref in this type

DISCRETE_OPAMP
- Fixed a bug (possible divide by 0)

DISCRETE_LFSR
- Fixed bug preventing output of shift reg to sub node 1

+++ src/emu/sound/disc_wav.c
@@ -479,7 +479,7 @@ static DISCRETE_STEP(dss_lfsr)
 
 		/* output the lfsr reg ?*/
 		if (context->out_lfsr_reg)
-			node->output[1] = context->lfsr_reg;
+			node->output[1] = (double) context->lfsr_reg;
 
 	}
 	if(!DSS_LFSR_NOISE__ENABLE)
